refactor: workspace settings
