+++ server/pkg/workspacesettings/workspacsettings.go
@@ -5,78 +5,10 @@ import (
 	"golang.org/x/exp/slices"
 )
 
-type ResourceType string
-
-const (
-	ResourceTypeTile    ResourceType = "TILE"
-	ResourceTypeTerrain ResourceType = "TERRAIN"
-)
-
-type TileType string
-
-const (
-	TileTypeDefault             TileType = "DEFAULT"
-	TileTypeLabelled            TileType = "LABELLED"
-	TileTypeRoadMap             TileType = "ROAD_MAP"
-	TileTypeStamenWaterColor    TileType = "STAMEN_WATERCOLOR"
-	TileTypeStamenToner         TileType = "STAMEN_TONER"
-	TileTypeOpenStreetMap       TileType = "OPEN_STREET_MAP"
-	TileTypeESRITopography      TileType = "ESRI_TOPOGRAPHY"
-	TileTypeEarthAtNight        TileType = "EARTH_AT_NIGHT"
-	TileTypeJapanGSIStandardMap TileType = "JAPAN_GSI_STANDARD_MAP"
-	TileTypeURL                 TileType = "URL"
-)
-
-type TerrainType string
-
-const (
-	TerrainTypeCesiumWorldTerrain TerrainType = "CESIUM_WORLD_TERRAIN"
-	TerrainTypeArcGISTerrain      TerrainType = "ARC_GIS_TERRAIN"
-	TerrainTypeCesiumIon          TerrainType = "CESIUM_ION"
-)
-
 type WorkspaceSettings struct {
 	id       ID
 	tiles    *ResourceList
 	terrains *ResourceList
-}
-
-type ResourceList struct {
-	resources        []*Resource
-	selectedResource *ResourceID
-	enabled          *bool
-}
-
-type Resource struct {
-	resourceType ResourceType
-	tile         *TileResource
-	terrain      *TerrainResource
-}
-
-type TileResource struct {
-	id    ResourceID
-	rtype TileType
-	props UrlResourceProps
-}
-
-type TerrainResource struct {
-	id    ResourceID
-	rtype TerrainType
-	props CesiumResourceProps
-}
-
-type UrlResourceProps struct {
-	name  string
-	url   string
-	image string
-}
-
-type CesiumResourceProps struct {
-	name                 string
-	url                  string
-	image                string
-	cesiumIonAssetId     string
-	cesiumIonAccessToken string
 }
 
 func (ws *WorkspaceSettings) ID() ID {
@@ -129,158 +61,4 @@ func (ws *WorkspaceSettings) Clone() *WorkspaceSettings {
 	}
 
 	return res
-}
-
-func (wr *ResourceList) Resources() []*Resource {
-	return slices.Clone(wr.resources)
-}
-
-func (wr *ResourceList) SelectedResource() *ResourceID {
-	return wr.selectedResource
-}
-
-func (wr *ResourceList) Enabled() *bool {
-	return wr.enabled
-}
-
-func (wr *ResourceList) SetResources(r []*Resource) {
-	wr.resources = slices.Clone(r)
-}
-
-func (wr *ResourceList) SetSelectedResource(rid *ResourceID) {
-	wr.selectedResource = rid.CloneRef()
-}
-
-func (wr *ResourceList) SetEnabled(s *bool) {
-	wr.enabled = util.CloneRef(s)
-}
-
-func NewResourceList(resources []*Resource, selectedResource *ResourceID, enabled *bool) *ResourceList {
-	return &ResourceList{
-		resources:        slices.Clone(resources),
-		selectedResource: util.CloneRef(selectedResource),
-		enabled:          util.CloneRef(enabled),
-	}
-}
-
-func (r *Resource) ResourceType() ResourceType {
-	return r.resourceType
-}
-
-func (r *Resource) Tile() *TileResource {
-	return r.tile
-}
-
-func (r *Resource) Terrain() *TerrainResource {
-	return r.terrain
-}
-
-func NewResource(resourceType ResourceType, tile *TileResource, terrain *TerrainResource) *Resource {
-	return &Resource{
-		resourceType: resourceType,
-		tile:         tile,
-		terrain:      terrain,
-	}
-}
-
-func (r *Resource) SetResourceType(n ResourceType) {
-	r.resourceType = n
-}
-
-func (r *Resource) SetTile(n *TileResource) {
-	r.tile = n
-}
-
-func (r *Resource) SetTerrain(n *TerrainResource) {
-	r.terrain = n
-}
-
-func NewTileResource(id ResourceID, rtype TileType, props UrlResourceProps) *TileResource {
-	return &TileResource{
-		id:    id.Clone(),
-		rtype: rtype,
-		props: props,
-	}
-}
-
-func (r *TileResource) ID() ResourceID {
-	return r.id
-}
-
-func (r *TileResource) Type() TileType {
-	return r.rtype
-}
-
-func (r *TileResource) Props() UrlResourceProps {
-	return r.props
-}
-
-func NewTerrainResource(id ResourceID, rtype TerrainType, props CesiumResourceProps) *TerrainResource {
-	return &TerrainResource{
-		id:    id.Clone(),
-		rtype: rtype,
-		props: props,
-	}
-}
-
-func (r *TerrainResource) ID() ResourceID {
-	return r.id
-}
-
-func (r *TerrainResource) Type() TerrainType {
-	return r.rtype
-}
-
-func (r *TerrainResource) Props() CesiumResourceProps {
-	return r.props
-}
-
-func NewURLResourceProps(name, url, image string) UrlResourceProps {
-	return UrlResourceProps{
-		name:  name,
-		url:   url,
-		image: image,
-	}
-}
-
-func (p UrlResourceProps) Name() string {
-	return p.name
-}
-
-func (p UrlResourceProps) URL() string {
-	return p.url
-}
-
-func (p UrlResourceProps) Image() string {
-	return p.image
-}
-
-func NewCesiumResourceProps(name, url, image, cesiumIonAssetId, cesiumIonAccessToken string) CesiumResourceProps {
-	return CesiumResourceProps{
-		name:                 name,
-		url:                  url,
-		image:                image,
-		cesiumIonAssetId:     cesiumIonAssetId,
-		cesiumIonAccessToken: cesiumIonAccessToken,
-	}
-}
-
-func (p CesiumResourceProps) Name() string {
-	return p.name
-}
-
-func (p CesiumResourceProps) URL() string {
-	return p.url
-}
-
-func (p CesiumResourceProps) Image() string {
-	return p.image
-}
-
-func (p CesiumResourceProps) CesiumIonAssetID() string {
-	return p.cesiumIonAssetId
-}
-
-func (p CesiumResourceProps) CesiumIonAccessToken() string {
-	return p.cesiumIonAccessToken
 }
